+++ babel.config.js
@@ -1,0 +1,5 @@
+// babel.config.js
+module.exports = {
+  presets: ["@babel/preset-env", "@babel/preset-react"],
+  plugins: ["@babel/plugin-proposal-class-properties"]
+};
