--- nwb.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-    type: "react-component",
-    npm: {
-        esModules: true,
-        umd: {
-            global: "ReactOffline",
-            externals: {
-                react: "React"
-            }
-        }
-    }
-};
