+++ package.json
@@ -21,22 +21,28 @@
     "url": "github.com/dashed"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/preset-env": "^7.4.3",
+    "@babel/preset-react": "^7.0.0",
     "babel-eslint": "^8.0.0",
+    "babel-jest": "^24.7.1",
     "eslint": "^4.6.1",
     "eslint-plugin-react": "^7.3.0",
+    "jest": "^24.7.1",
     "microbundle": "^0.11.0",
     "prettier": "^1.6.1",
     "react": "^16.0.0",
-    "react-dom": "^16.0.0"
+    "react-dom": "^16.0.0",
+    "react-test-renderer": "^16.8.6"
   },
   "scripts": {
-    "pretty": "prettier --write --tab-width 4 'src/**/*.js' 'test/**/*.js'",
+    "pretty": "prettier --write 'src/**/*.js' 'test/**/*.js' './*.js'",
     "lint": "eslint src test",
     "build": "microbundle --no-compress --external=all --sourcemap=false",
     "dev": "microbundle watch --no-compress  --external=all --sourcemap=false",
-    "test": "nwb test-react",
+    "test": "jest test/",
     "travis": "npm run lint && npm run test",
-    "prepublish": "npm run lint && npm run test && npm run build"
+    "prepare": "npm run lint && npm run test && npm run build"
   },
   "eslintConfig": {
     "parser": "babel-eslint",
@@ -51,18 +57,6 @@
     "extends": [
       "eslint:recommended",
       "plugin:react/recommended"
-    ]
-  },
-  "babel": {
-    "presets": [
-      [
-        "env",
-        {
-          "targets": {
-            "node": "0.12"
-          }
-        }
-      ]
     ]
   },
   "keywords": [
